performs splitting into controls
