+++ gui/src/controls/mod.rs
@@ -1,0 +1,4 @@
+mod tweak_control;
+mod next_round_control;
+mod game_status_display;
+mod resource_control;
